First working version

+++ app/src/main/java/com/example/weatherapp/data/model/forecast/ForecastWeather.kt
@@ -12,7 +12,7 @@ data class ForecastWeather(
     val clouds: Clouds,
     val wind: Wind,
     val visibility: Int,
-    @SerializedName("pop") val probabilityOfPrecipitation: Int,
+    @SerializedName("pop") val probabilityOfPrecipitation: Double,
     val rain: ForecastRain?,
     val snow: ForecastSnow?,
     val sys: ForecastSys,
